+++ halo-third-party-sdk-model/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.0]
+### Features
+* Activity data object was updated to include Activity Summary, and Activity Sessions now support intensity durations.
+* Sleep data object was updated to include session start and end times.
+* Notification type was updated to include 'HEALTH_METRIC_DELETION' and 'HEALTH_METRIC_DELETION_TEST'.
+
 ## [1.0.0]
 ### Features
 * Initial release of Halo Third Party SDK.